refactor: rename file names to kebab-case
