--- app/helpers/ui/colorSchemaInit.entry.ts
@@ -1,3 +1,0 @@
-import { initColorScheme } from "./colorSchema"
-
-initColorScheme()
