+++ app/hooks/useColorScheme.ts
@@ -8,7 +8,7 @@ import {
   resolveColorScheme,
   setStoredColorScheme,
   THEME_STORAGE_KEY,
-} from "../helpers/ui/colorSchema"
+} from "../helpers/ui/color-scheme"
 
 const COLOR_SCHEME_EVENT = "color-scheme:change"
